+++ tests/Unit/Dto/Order/FulfillmentOrder/GetListTest.php
@@ -44,10 +44,30 @@ class GetListTest extends TestCase
                     'currencyCode' => 'USD',
                     'notes' => 'Some note',
                     'specialInstructions' => '',
-                    "orderSubmittedDate" => "2019-09-04T07:26:42.535Z",
-                    "acceptedDate" => "2019-09-04T07:26:42.535Z",
-                    "readyDate" => "2019-09-04T07:26:42.535Z",
-                    "completedDate" => "2019-09-04T07:26:42.535Z"
+                    'submitDate' => "2019-09-04T07:26:42.535Z",
+                    'acceptDate' => "2019-09-04T07:26:42.535Z",
+                    'readyDate' => "2019-09-04T07:26:42.535Z",
+                    'completeDate' => "2019-09-04T07:26:42.535Z",
+                    'fulfillmentOrderAddress' => [
+                        "orderIndex" => "123",
+                        "type" => "pickup",
+                        "firstName" => "John",
+                        "middleName" => "D.",
+                        "lastName" => "Doe",
+                        "company" => "Shopgate Inc",
+                        "address1" => "12 Somestreet",
+                        "address2" => "Unknown Type: string,null",
+                        "address3" => "Unknown Type: string,null",
+                        "address4" => "Unknown Type: string,null",
+                        "city" => "Austin",
+                        "region" => "TX",
+                        "postalCode" => '78732',
+                        "country" => "US",
+                        "phone" => "+1 000-000-0000",
+                        "fax" => "+1 000-000-0000",
+                        "mobile" => "+1 000-000-0000",
+                        "emailAddress" => "john@doe.com"
+                    ],
                 ]
             ]
         ];
@@ -70,7 +90,10 @@ class GetListTest extends TestCase
         $this->assertEquals($entry['fulfillmentOrders'][0]['locationCode'], $get->getLocationCode());
         $this->assertEquals($entry['fulfillmentOrders'][0]['type'], $get->getType());
         $this->assertEquals($entry['fulfillmentOrders'][0]['customerId'], $get->getCustomerId());
-        $this->assertEquals($entry['fulfillmentOrders'][0]['externalCustomerNumber'], $get->getExternalCustomerNumber());
+        $this->assertEquals(
+            $entry['fulfillmentOrders'][0]['externalCustomerNumber'],
+            $get->getExternalCustomerNumber()
+        );
         $this->assertEquals($entry['fulfillmentOrders'][0]['routeType'], $get->getRouteType());
         $this->assertEquals($entry['fulfillmentOrders'][0]['expedited'], $get->getExpedited());
         $this->assertEquals($entry['fulfillmentOrders'][0]['status'], $get->getStatus());
@@ -82,5 +105,31 @@ class GetListTest extends TestCase
         $this->assertEquals($entry['fulfillmentOrders'][0]['currencyCode'], $get->getCurrencyCode());
         $this->assertEquals($entry['fulfillmentOrders'][0]['notes'], $get->getNotes());
         $this->assertEquals($entry['fulfillmentOrders'][0]['specialInstructions'], $get->getSpecialInstructions());
+        $this->assertEquals($entry['fulfillmentOrders'][0]['submitDate'], $get->getSubmitDate());
+        $this->assertEquals($entry['fulfillmentOrders'][0]['acceptDate'], $get->getAcceptDate());
+        $this->assertEquals($entry['fulfillmentOrders'][0]['readyDate'], $get->getReadyDate());
+        $this->assertEquals($entry['fulfillmentOrders'][0]['completeDate'], $get->getCompleteDate());
+
+        $fulfillmentOrderAddress = $get->getFulfillmentOrderAddress();
+        $entryAddress = $entry['fulfillmentOrders'][0]['fulfillmentOrderAddress'];
+        $this->assertInstanceOf(FulfillmentOrder\Dto\FulfillmentOrderAddress::class, $fulfillmentOrderAddress);
+        $this->assertEquals($entryAddress['orderIndex'], $fulfillmentOrderAddress->getOrderIndex());
+        $this->assertEquals($entryAddress['type'], $fulfillmentOrderAddress->getType());
+        $this->assertEquals($entryAddress['firstName'], $fulfillmentOrderAddress->getFirstName());
+        $this->assertEquals($entryAddress['middleName'], $fulfillmentOrderAddress->getMiddleName());
+        $this->assertEquals($entryAddress['lastName'], $fulfillmentOrderAddress->getLastName());
+        $this->assertEquals($entryAddress['company'], $fulfillmentOrderAddress->getCompany());
+        $this->assertEquals($entryAddress['address1'], $fulfillmentOrderAddress->getAddress1());
+        $this->assertEquals($entryAddress['address2'], $fulfillmentOrderAddress->getAddress2());
+        $this->assertEquals($entryAddress['address3'], $fulfillmentOrderAddress->getAddress3());
+        $this->assertEquals($entryAddress['address4'], $fulfillmentOrderAddress->getAddress4());
+        $this->assertEquals($entryAddress['city'], $fulfillmentOrderAddress->getCity());
+        $this->assertEquals($entryAddress['region'], $fulfillmentOrderAddress->getRegion());
+        $this->assertEquals($entryAddress['postalCode'], $fulfillmentOrderAddress->getPostalCode());
+        $this->assertEquals($entryAddress['country'], $fulfillmentOrderAddress->getCountry());
+        $this->assertEquals($entryAddress['phone'], $fulfillmentOrderAddress->getPhone());
+        $this->assertEquals($entryAddress['fax'], $fulfillmentOrderAddress->getFax());
+        $this->assertEquals($entryAddress['mobile'], $fulfillmentOrderAddress->getMobile());
+        $this->assertEquals($entryAddress['emailAddress'], $fulfillmentOrderAddress->getEmailAddress());
     }
 }